+++ plugin-gradle/src/main/java/com/diffplug/gradle/spotless/SpotlessTask.java
@@ -18,8 +18,6 @@ package com.diffplug.gradle.spotless;
 import java.io.File;
 import java.io.Serializable;
 import java.nio.charset.Charset;
-import java.nio.file.Files;
-import java.nio.file.StandardOpenOption;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -47,7 +45,6 @@ import com.diffplug.spotless.Formatter;
 import com.diffplug.spotless.FormatterStep;
 import com.diffplug.spotless.LineEnding;
 import com.diffplug.spotless.PaddedCell;
-import com.diffplug.spotless.PaddedCellBulk;
 import com.diffplug.spotless.extra.integration.DiffMessageFormatter;
 
 public class SpotlessTask extends DefaultTask {
@@ -74,16 +71,15 @@ public class SpotlessTask extends DefaultTask {
 		this.lineEndingsPolicy = Objects.requireNonNull(lineEndingsPolicy);
 	}
 
-	// set by FormatExtension
-	protected boolean paddedCell = false;
-
-	@Input
+	@Deprecated
+	@Internal
 	public boolean isPaddedCell() {
-		return paddedCell;
+		return true;
 	}
 
+	@Deprecated
 	public void setPaddedCell(boolean paddedCell) {
-		this.paddedCell = paddedCell;
+		getLogger().warn("PaddedCell is now always on, and cannot be turned off.");
 	}
 
 	protected String filePatterns = "";
@@ -228,14 +224,7 @@ public class SpotlessTask extends DefaultTask {
 		}
 
 		// create the formatter
-		try (Formatter formatter = Formatter.builder()
-				.lineEndingsPolicy(lineEndingsPolicy)
-				.encoding(Charset.forName(encoding))
-				.rootDir(getProject().getRootDir().toPath())
-				.steps(steps)
-				.exceptionPolicy(exceptionPolicy)
-				.build()) {
-
+		try (Formatter formatter = buildFormatter()) {
 			if (apply) {
 				List<File> changedFiles = applyAnyChanged(formatter, outOfDate);
 				if (!changedFiles.isEmpty()) {
@@ -254,6 +243,16 @@ public class SpotlessTask extends DefaultTask {
 		}
 	}
 
+	Formatter buildFormatter() {
+		return Formatter.builder()
+				.lineEndingsPolicy(lineEndingsPolicy)
+				.encoding(Charset.forName(encoding))
+				.rootDir(getProject().getRootDir().toPath())
+				.steps(steps)
+				.exceptionPolicy(exceptionPolicy)
+				.build();
+	}
+
 	static class LastApply implements Serializable {
 		private static final long serialVersionUID = 6245070824310295090L;
 
@@ -263,42 +262,16 @@ public class SpotlessTask extends DefaultTask {
 
 	private List<File> applyAnyChanged(Formatter formatter, List<File> outOfDate) throws Exception {
 		List<File> changed = new ArrayList<>(outOfDate.size());
-		if (isPaddedCell()) {
-			for (File file : outOfDate) {
-				getLogger().debug("Applying format to " + file);
-				if (PaddedCellBulk.applyAnyChanged(formatter, file)) {
-					changed.add(file);
-				}
-			}
-		} else {
-			boolean anyMisbehave = false;
-			for (File file : outOfDate) {
-				getLogger().debug("Applying format to " + file);
-				String unixResultIfDirty = formatter.applyToAndReturnResultIfDirty(file);
-				if (unixResultIfDirty != null) {
-					changed.add(file);
-				}
-				// because apply will count as up-to-date, it's important
-				// that every call to apply will get a PaddedCell check
-				if (!anyMisbehave && unixResultIfDirty != null) {
-					String onceMore = formatter.compute(unixResultIfDirty, file);
-					//  f(f(input) == f(input) for an idempotent function
-					if (!onceMore.equals(unixResultIfDirty)) {
-						// it's not idempotent.  but, if it converges, then it's likely a glitch that won't reoccur,
-						// so there's no need to make a bunch of noise for the user
-						PaddedCell result = PaddedCell.check(formatter, file, onceMore);
-						if (result.type() == PaddedCell.Type.CONVERGE) {
-							String finalResult = formatter.computeLineEndings(result.canonical(), file);
-							Files.write(file.toPath(), finalResult.getBytes(formatter.getEncoding()), StandardOpenOption.TRUNCATE_EXISTING);
-						} else {
-							// it didn't converge, so the user is going to need padded cell mode
-							anyMisbehave = true;
-						}
-					}
-				}
-			}
-			if (anyMisbehave) {
-				throw PaddedCellGradle.youShouldTurnOnPaddedCell(this);
+		for (File file : outOfDate) {
+			getLogger().debug("Applying format to " + file);
+			PaddedCell.DirtyState dirtyState = PaddedCell.calculateDirtyState(formatter, file);
+			if (dirtyState.isClean()) {
+				// do nothing
+			} else if (dirtyState.didNotConverge()) {
+				getLogger().warn("Skipping '" + file + "' because it does not converge.  Run `spotlessDiagnose` to understand why");
+			} else {
+				dirtyState.writeCanonicalTo(file);
+				changed.add(file);
 			}
 		}
 		return changed;
@@ -308,21 +281,17 @@ public class SpotlessTask extends DefaultTask {
 		List<File> problemFiles = new ArrayList<>();
 		for (File file : outOfDate) {
 			getLogger().debug("Checking format on " + file);
-			if (!formatter.isClean(file)) {
+			PaddedCell.DirtyState dirtyState = PaddedCell.calculateDirtyState(formatter, file);
+			if (dirtyState.isClean()) {
+				// do nothing
+			} else if (dirtyState.didNotConverge()) {
+				getLogger().warn("Skipping '" + file + "' because it does not converge.  Run `spotlessDiagnose` to understand why");
+			} else {
 				problemFiles.add(file);
 			}
 		}
-		if (paddedCell) {
-			PaddedCellGradle.check(this, formatter, problemFiles);
-		} else {
-			if (!problemFiles.isEmpty()) {
-				// if we're not in paddedCell mode, we'll check if maybe we should be
-				if (PaddedCellBulk.anyMisbehave(formatter, problemFiles)) {
-					throw PaddedCellGradle.youShouldTurnOnPaddedCell(this);
-				} else {
-					throw formatViolationsFor(formatter, problemFiles);
-				}
-			}
+		if (!problemFiles.isEmpty()) {
+			throw formatViolationsFor(formatter, problemFiles);
 		}
 	}
 
@@ -330,7 +299,6 @@ public class SpotlessTask extends DefaultTask {
 	GradleException formatViolationsFor(Formatter formatter, List<File> problemFiles) {
 		return new GradleException(DiffMessageFormatter.builder()
 				.runToFix("Run 'gradlew spotlessApply' to fix these violations.")
-				.isPaddedCell(paddedCell)
 				.formatter(formatter)
 				.problemFiles(problemFiles)
 				.getMessage());