+++ plugin-gradle/src/test/java/com/diffplug/gradle/spotless/PaddedCellTaskTest.java
@@ -23,7 +23,6 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.Locale;
 
-import org.assertj.core.api.Assertions;
 import org.gradle.api.Project;
 import org.junit.Assert;
 import org.junit.Test;
@@ -44,12 +43,14 @@ public class PaddedCellTaskTest extends ResourceHarness {
 	}
 
 	private class Bundle {
+		String name;
 		Project project = TestProvisioner.gradleProject(rootFolder());
 		File file;
 		SpotlessTask check;
 		SpotlessTask apply;
 
 		Bundle(String name, FormatterFunc function) throws IOException {
+			this.name = name;
 			file = setFile("src/test." + name).toContent("CCC");
 			FormatterStep step = FormatterStep.createNeverUpToDate(name, function);
 			check = createCheckTask(name, step);
@@ -76,12 +77,6 @@ public class PaddedCellTaskTest extends ResourceHarness {
 			return task;
 		}
 
-		private Bundle paddedCell() {
-			check.setPaddedCell(true);
-			apply.setPaddedCell(true);
-			return this;
-		}
-
 		private String checkFailureMsg() {
 			try {
 				execute(check);
@@ -89,6 +84,12 @@ public class PaddedCellTaskTest extends ResourceHarness {
 			} catch (Exception e) {
 				return e.getMessage();
 			}
+		}
+
+		private void diagnose() throws IOException {
+			SpotlessDiagnoseTask diagnose = project.getTasks().create("spotless" + SpotlessPlugin.capitalize(name) + "Diagnose", SpotlessDiagnoseTask.class);
+			diagnose.source = check;
+			diagnose.performAction();
 		}
 	}
 
@@ -109,19 +110,11 @@ public class PaddedCellTaskTest extends ResourceHarness {
 	}
 
 	@Test
-	public void failsWithoutPaddedCell() throws IOException {
-		Assertions.assertThat(wellbehaved().checkFailureMsg()).startsWith("The following files had format violations");
-		Assertions.assertThat(cycle().checkFailureMsg()).startsWith("You have a misbehaving rule");
-		Assertions.assertThat(converge().checkFailureMsg()).startsWith("You have a misbehaving rule");
-		Assertions.assertThat(diverge().checkFailureMsg()).startsWith("You have a misbehaving rule");
-	}
-
-	@Test
 	public void paddedCellApply() throws Exception {
-		Bundle wellbehaved = wellbehaved().paddedCell();
-		Bundle cycle = cycle().paddedCell();
-		Bundle converge = converge().paddedCell();
-		Bundle diverge = diverge().paddedCell();
+		Bundle wellbehaved = wellbehaved();
+		Bundle cycle = cycle();
+		Bundle converge = converge();
+		Bundle diverge = diverge();
 
 		execute(wellbehaved.apply);
 		execute(cycle.apply);
@@ -140,11 +133,11 @@ public class PaddedCellTaskTest extends ResourceHarness {
 	}
 
 	@Test
-	public void paddedCellCheckFailureFiles() throws Exception {
-		wellbehaved().paddedCell().checkFailureMsg();
-		cycle().paddedCell().checkFailureMsg();
-		converge().paddedCell().checkFailureMsg();
-		execute(diverge().paddedCell().check);
+	public void diagnose() throws Exception {
+		wellbehaved().diagnose();
+		cycle().diagnose();
+		converge().diagnose();
+		diverge().diagnose();
 
 		assertFolderContents("build",
 				"spotless-diagnose-converge",
@@ -179,7 +172,7 @@ public class PaddedCellTaskTest extends ResourceHarness {
 
 	@Test
 	public void paddedCellCheckCycleFailureMsg() throws IOException {
-		assertFailureMessage(cycle().paddedCell(),
+		assertFailureMessage(cycle(),
 				"The following files had format violations:",
 				slashify("    src/test.cycle"),
 				"        @@ -1 +1 @@",
@@ -190,7 +183,7 @@ public class PaddedCellTaskTest extends ResourceHarness {
 
 	@Test
 	public void paddedCellCheckConvergeFailureMsg() throws IOException {
-		assertFailureMessage(converge().paddedCell(),
+		assertFailureMessage(converge(),
 				"The following files had format violations:",
 				slashify("    src/test.converge"),
 				"        @@ -1 +0,0 @@",