+++ CHANGES.md
@@ -11,7 +11,7 @@ We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format (
 
 ## [Unreleased]
 ### Added
-* Support for google-java-format 1.8. ([#562](https://github.com/diffplug/spotless/issues/562))
+* Support for google-java-format 1.8 (including test infrastructure for Java 11). ([#562](https://github.com/diffplug/spotless/issues/562))
 
 ## [1.28.1] - 2020-04-02
 ### Fixed