+++ CHANGES.md
@@ -13,6 +13,11 @@ We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format (
 ### Added
 * Support for google-java-format 1.8 (including test infrastructure for Java 11). ([#562](https://github.com/diffplug/spotless/issues/562))
 * Improved PaddedCell such that it is as performant as non-padded cell - no reason not to have it always enabled.  Deprecated all of `PaddedCellBulk`. ([#561](https://github.com/diffplug/spotless/pull/561))
+### Changed
+* Updated a bunch of dependencies, most notably: ([#564](https://github.com/diffplug/spotless/pull/564))
+  * jgit `5.5.0.201909110433-r` -> `5.7.0.202003110725-r`
+  * gradle `6.2.2` -> `6.3`
+  * spotbugs gradle plugin `2.0.0` -> `4.0.8`
 
 ## [1.28.1] - 2020-04-02
 ### Fixed